move ns
